requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
matplotlib
numpy
jinja2
scikit-learn
seaborn
